+++ selfdrive/car/volkswagen/interface.py
@@ -1,18 +1,16 @@
-from cereal import car, log
-from common.realtime import sec_since_boot
+from cereal import car
 from selfdrive.config import Conversions as CV
-from selfdrive.controls.lib.drive_helpers import create_event, get_events, EventTypes as ET
+from selfdrive.controls.lib.drive_helpers import create_event, EventTypes as ET
 from selfdrive.controls.lib.vehicle_model import VehicleModel
-from selfdrive.car.volkswagen.values import CAR, FINGERPRINTS, ECU_FINGERPRINT, ECU, gra_acc_buttons_dict
+from selfdrive.car.volkswagen.values import CAR, BUTTON_STATES
 from selfdrive.car.volkswagen.carstate import CarState, get_mqb_gateway_can_parser, get_mqb_extended_can_parser
 from common.params import Params
-from selfdrive.car import STD_CARGO_KG, scale_rot_inertia, scale_tire_stiffness, gen_empty_fingerprint, is_ecu_disconnected
+from selfdrive.car import STD_CARGO_KG, scale_rot_inertia, scale_tire_stiffness, gen_empty_fingerprint
 from selfdrive.car.interfaces import CarInterfaceBase
 
-class CanBus(object):
-  def __init__(self):
-    self.gateway = 0
-    self.extended = 2
+class CANBUS:
+  gateway = 0
+  extended = 2
 
 class CarInterface(CarInterfaceBase):
   def __init__(self, CP, CarController):
@@ -23,23 +21,19 @@ class CarInterface(CarInterfaceBase):
 
     self.gasPressedPrev = False
     self.brakePressedPrev = False
-    self.rightBlinkerPrev = False
-    self.leftBlinkerPrev = False
     self.cruiseStateEnabledPrev = False
     self.displayMetricUnitsPrev = None
-    self.gra_acc_buttons_prev = gra_acc_buttons_dict.copy()
+    self.buttonStatesPrev = BUTTON_STATES.copy()
 
     # *** init the major players ***
-    canbus = CanBus()
-    self.CS = CarState(CP, canbus)
+    self.CS = CarState(CP, CANBUS)
     self.VM = VehicleModel(CP)
-    self.gw_cp = get_mqb_gateway_can_parser(CP, canbus)
-    self.ex_cp = get_mqb_extended_can_parser(CP, canbus)
+    self.gw_cp = get_mqb_gateway_can_parser(CP, CANBUS)
+    self.ex_cp = get_mqb_extended_can_parser(CP, CANBUS)
 
     # sending if read only is False
     if CarController is not None:
-      self.CC = CarController(canbus, CP.carFingerprint)
-
+      self.CC = CarController(CANBUS, CP.carFingerprint)
 
   @staticmethod
   def compute_gb(accel, speed):
@@ -48,9 +42,6 @@ class CarInterface(CarInterfaceBase):
   @staticmethod
   def get_params(candidate, fingerprint=gen_empty_fingerprint(), vin="", has_relay=False):
     ret = car.CarParams.new_message()
-
-    # FIXME: Temp hack while working out fingerprint side
-    has_auto_trans = True
 
     ret.carFingerprint = candidate
     ret.isPandaBlack = has_relay
@@ -95,11 +86,7 @@ class CarInterface(CarInterfaceBase):
       ret.lateralTuning.pid.kiV = [0.20, 0.10, 0.10, 0.05, 0.05]
       tire_stiffness_factor = 0.6
 
-    # FIXME: Need to find a clean way to handle e-Golf without Getriebe_11 message
-    if has_auto_trans:
-      ret.transmissionType = car.CarParams.TransmissionType.automatic
-    else:
-      ret.transmissionType = car.CarParams.TransmissionType.manual
+    ret.transmissionType = car.CarParams.TransmissionType.automatic
 
     # FIXME: follow 0.6.5 Comma refactoring to ensure camera-side is detected okay
     # ret.enableCamera = is_ecu_disconnected(fingerprint[0], FINGERPRINTS, ECU_FINGERPRINT, candidate, ECU.CAM) or has_relay
@@ -108,17 +95,15 @@ class CarInterface(CarInterfaceBase):
 
     # No support for OP longitudinal control on Volkswagen at this time.
     ret.gasMaxBP = [0.]
-    ret.gasMaxV = [.5]
+    ret.gasMaxV = [0.]
     ret.brakeMaxBP = [0.]
-    ret.brakeMaxV = [1.]
+    ret.brakeMaxV = [0.]
     ret.longitudinalTuning.deadzoneBP = [0.]
     ret.longitudinalTuning.deadzoneV = [0.]
-    ret.longitudinalTuning.kpBP = [5., 35.]
-    ret.longitudinalTuning.kpV = [2.4, 1.5]
+    ret.longitudinalTuning.kpBP = [0.]
+    ret.longitudinalTuning.kpV = [0.]
     ret.longitudinalTuning.kiBP = [0.]
-    ret.longitudinalTuning.kiV = [0.36]
-    ret.stoppingControl = True
-    ret.startAccel = 0.8
+    ret.longitudinalTuning.kiV = [0.]
 
     # TODO: get actual value, for now starting with reasonable value for
     # civic and scaling by mass and wheelbase
@@ -185,63 +170,21 @@ class CarInterface(CarInterfaceBase):
       params.put("IsMetric", "1" if self.CS.displayMetricUnits else "0")
 
     # Blinker switch updates
-    ret.leftBlinker = self.CS.leftBlinker
-    ret.rightBlinker = self.CS.rightBlinker
-
-    if ret.leftBlinker != self.leftBlinkerPrev:
-      be = car.CarState.ButtonEvent.new_message()
-      be.type = 'leftBlinker'
-      be.pressed = ret.leftBlinker
-      buttonEvents.append(be)
-
-    if ret.rightBlinker != self.rightBlinkerPrev:
-      be = car.CarState.ButtonEvent.new_message()
-      be.type = 'rightBlinker'
-      be.pressed = ret.rightBlinker
-      buttonEvents.append(be)
+    ret.leftBlinker = self.CS.buttonStates["leftBlinker"]
+    ret.rightBlinker = self.CS.buttonStates["rightBlinker"]
 
     # ACC cruise state
     ret.cruiseState.available = self.CS.accAvailable
     ret.cruiseState.enabled = self.CS.accEnabled
     ret.cruiseState.speed = self.CS.accSetSpeed
 
-    # Process button press or release events from ACC steering wheel or
-    # control stalk buttons.
-    if self.CS.gra_acc_buttons != self.gra_acc_buttons_prev:
-      if self.CS.gra_acc_buttons["main"] != self.gra_acc_buttons_prev["main"]:
+    # Check for and process state-change events (button press or release) from
+    # the turn stalk switch or ACC steering wheel/control stalk buttons.
+    for button in self.CS.buttonStates:
+      if self.CS.buttonStates[button] != self.buttonStatesPrev[button]:
         be = car.CarState.ButtonEvent.new_message()
-        be.type = 'altButton3'
-        be.pressed = bool(self.CS.gra_acc_buttons["main"])
-        buttonEvents.append(be)
-      if self.CS.gra_acc_buttons["set"] != self.gra_acc_buttons_prev["set"]:
-        be = car.CarState.ButtonEvent.new_message()
-        be.type = 'setCruise'
-        be.pressed = bool(self.CS.gra_acc_buttons["set"])
-        buttonEvents.append(be)
-      if self.CS.gra_acc_buttons["resume"] != self.gra_acc_buttons_prev["resume"]:
-        be = car.CarState.ButtonEvent.new_message()
-        be.type = 'resumeCruise'
-        be.pressed = bool(self.CS.gra_acc_buttons["resume"])
-        buttonEvents.append(be)
-      if self.CS.gra_acc_buttons["cancel"] != self.gra_acc_buttons_prev["cancel"]:
-        be = car.CarState.ButtonEvent.new_message()
-        be.type = 'cancel'
-        be.pressed = bool(self.CS.gra_acc_buttons["cancel"])
-        buttonEvents.append(be)
-      if self.CS.gra_acc_buttons["accel"] != self.gra_acc_buttons_prev["accel"]:
-        be = car.CarState.ButtonEvent.new_message()
-        be.type = 'accelCruise'
-        be.pressed = bool(self.CS.gra_acc_buttons["accel"])
-        buttonEvents.append(be)
-      if self.CS.gra_acc_buttons["decel"] != self.gra_acc_buttons_prev["decel"]:
-        be = car.CarState.ButtonEvent.new_message()
-        be.type = 'decelCruise'
-        be.pressed = bool(self.CS.gra_acc_buttons["decel"])
-        buttonEvents.append(be)
-      if self.CS.gra_acc_buttons["timegap"] != self.gra_acc_buttons_prev["timegap"]:
-        be = car.CarState.ButtonEvent.new_message()
-        be.type = 'gapAdjustCruise'
-        be.pressed = bool(self.CS.gra_acc_buttons["timegap"])
+        be.type = button
+        be.pressed = self.CS.buttonStates[button]
         buttonEvents.append(be)
 
     # Vehicle operation safety checks and events
@@ -253,9 +196,6 @@ class CarInterface(CarInterfaceBase):
       events.append(create_event('reverseGear', [ET.NO_ENTRY, ET.IMMEDIATE_DISABLE]))
     if not ret.gearShifter == 'drive' and not ret.gearShifter == 'eco':
       events.append(create_event('wrongGear', [ET.NO_ENTRY, ET.SOFT_DISABLE]))
-    # FIXME: (ab)using wrongGear until we have a cereal event for clutchPressed
-    if ret.clutchPressed:
-      events.append(create_event('wrongGear', [ET.NO_ENTRY]))
     if self.CS.stabilityControlDisabled:
       events.append(create_event('espDisabled', [ET.NO_ENTRY, ET.SOFT_DISABLE]))
     if self.CS.parkingBrakeSet:
@@ -265,7 +205,7 @@ class CarInterface(CarInterfaceBase):
     if self.CS.accFault:
       events.append(create_event('radarFault', [ET.NO_ENTRY, ET.IMMEDIATE_DISABLE]))
     if self.CS.steeringFault:
-      events.append(create_event('steerTempUnavailable', [ET.NO_ENTRY, ET.IMMEDIATE_DISABLE]))
+      events.append(create_event('steerTempUnavailable', [ET.NO_ENTRY, ET.WARNING]))
 
     # Per the Comma safety model, disable on pedals rising edge or when brake
     # is pressed and speed isn't zero.
@@ -290,11 +230,9 @@ class CarInterface(CarInterfaceBase):
     # update previous car states
     self.gasPressedPrev = ret.gasPressed
     self.brakePressedPrev = ret.brakePressed
-    self.leftBlinkerPrev = ret.leftBlinker
-    self.rightBlinkerPrev = ret.rightBlinker
     self.cruiseStateEnabledPrev = ret.cruiseState.enabled
     self.displayMetricUnitsPrev = self.CS.displayMetricUnits
-    self.gra_acc_buttons_prev = self.CS.gra_acc_buttons.copy()
+    self.buttonStatesPrev = self.CS.buttonStates.copy()
 
     # cast to reader so it can't be modified
     return ret.as_reader()